Bugfix: reductions for cancelled computations
Background
----------

The computation graph can *request* and *cancel* computations. It requests
computations when thunks become executable (either initially or after thunks
resolve to values). It cancels computations when future-depending thunks
resolve, if this yields their future-dependencies (globally) unneeded.

Bug
---

When the computation graph
   1. Requests a computation
   2. Cancels it
   3. Recieves a reduction for that computation
it error'd out because it could not find the computation anymore.

Fix
---

Check for this case, and drop the reduction.

Alternatives
------------

Have the reductor track which computations are cancelled, and not submit
reductions for cancelled computations.

I rejected this because (currently) the graph knows what is needed, and the
reductor doesn't.